// src/types/CustomerType.ts
interface CustomerType {
  // role: string;
  firstName: string;
  lastName: string;
  email: string;
  phoneNumber: string;
  companyName: string;
  address1: string;
  address2: string;
  city: string;
  state: string;
  country: string;
  zipCode: string;
  notes: string;
}
export default CustomerType;